back-end finalizado, fron-end prototipo

+++ crud/src/main/java/com/partricioturismo/crud/repositories/OnibusRepository.java
@@ -3,5 +3,5 @@ package com.partricioturismo.crud.repositories;
 import com.partricioturismo.crud.model.Onibus;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OnibusRepository extends JpaRepository<Onibus, Integer> {
+public interface OnibusRepository extends JpaRepository<Onibus, Long> {
 }
