+++ crud/src/main/java/com/partricioturismo/crud/repositories/ViagemRepository.java
@@ -3,5 +3,5 @@ package com.partricioturismo.crud.repositories;
 import com.partricioturismo.crud.model.Viagem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ViagemRepository extends JpaRepository<Viagem, Integer> {
+public interface ViagemRepository extends JpaRepository<Viagem, Long> {
 }
